nca_trading_bot/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.5.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data fetching and processing
yfinance>=0.2.0
alpaca-trade-api>=3.0.0
pandas-datareader>=0.10.0

# Technical analysis
talib-binary>=0.4.19
ta-lib>=0.4.25

# Machine learning and reinforcement learning
gymnasium>=0.29.0
stable-baselines3>=2.0.0
scikit-learn>=1.3.0

# Scientific computing
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Financial libraries
quantlib>=1.30
pyfolio>=0.9.2
zipline>=2.0.0

# Data processing and manipulation
polars>=0.19.0
dask>=2023.9.0
pyarrow>=13.0.0

# API and web frameworks
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0

# Configuration and CLI
click>=8.1.0
python-dotenv>=1.0.0
rich>=13.5.0

# Logging and monitoring
tensorboard>=2.14.0
wandb>=0.15.0
structlog>=23.1.0

# Performance and optimization
numba>=0.58.0
cython>=3.0.0
psutil>=5.9.0
gputil>=1.4.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0
mkdocs>=1.5.0

# Jupyter and interactive computing
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0
jupyterlab>=4.0.0

# Async and concurrency
aiohttp>=3.8.0
httpx>=0.24.0
asyncio>=3.4.3

# Image processing (for charts and visualization)
pillow>=10.0.0
opencv-python>=4.8.0

# Serialization and storage
h5py>=3.9.0
pickle5>=0.0.12
dill>=0.3.7

# Cryptography and security
cryptography>=41.0.0
bcrypt>=4.0.0

# System and OS
pyyaml>=6.0.1
toml>=0.10.2
configparser>=6.0.0

# Optional: GPU acceleration
cupy-cuda12x>=12.2.0

# Optional: QuantConnect integration
quantconnect-api>=1.0.0

# Optional: Interactive Brokers integration
ibapi>=10.19.0

# Optional: MetaTrader integration
metatrader5>=5.0.45

# Optional: TradingView integration
tradingview-ta>=3.3.0

# Optional: Bloomberg integration
# blpapi>=3.20.0  # Requires Bloomberg Terminal

# Optional: Quandl integration
# quandl>=3.7.0  # Requires API key

# Optional: Alpha Vantage integration
# alpha-vantage>=2.3.1  # Requires API key

# Optional: Polygon.io integration
# polygon-api-client>=1.10.0  # Requires API key

# Optional: Financial Modeling Prep integration
# fmp-api>=1.0.0  # Requires API key

# Optional: Yahoo Finance alternatives
# yahoo-finance>=1.4.0
# yahooquery>=2.3.0

# Optional: News and sentiment analysis
# newsapi-python>=0.2.7  # Requires API key
# textblob>=0.17.1
# vaderSentiment>=3.3.2
# transformers>=4.21.0
# torchtext>=0.15.0

# Optional: Database integrations
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0
# pymongo>=4.5.0
# redis>=5.0.0
# influxdb-client>=1.36.0

# Optional: Cloud storage
# boto3>=1.28.0
# google-cloud-storage>=2.10.0
# azure-storage-blob>=12.17.0

# Optional: Message queues
# pika>=1.3.0  # RabbitMQ
# confluent-kafka>=2.2.0
# redis>=5.0.0

# Optional: Monitoring and alerting
# prometheus-client>=0.17.0
# sentry-sdk>=1.30.0
# datadog>=0.48.0

# Optional: Web scraping
# beautifulsoup4>=4.12.0
# selenium>=4.14.0
# scrapy>=2.11.0

# Optional: PDF processing
# pypdf2>=3.0.0
# reportlab>=4.0.0

# Optional: Excel processing
# openpyxl>=3.1.0
# xlrd>=2.0.1
# xlsxwriter>=3.1.0

# Optional: Email integration
# smtplib
# email-validator>=2.0.0

# Optional: SMS integration
# twilio>=8.5.0

# Optional: Slack integration
# slack-sdk>=3.22.0

# Optional: Discord integration
# discord.py>=2.3.0

# Optional: Telegram integration
# python-telegram-bot>=20.5.0

# Optional: Docker integration
# docker>=6.1.0
# docker-compose>=1.29.0

# Optional: Kubernetes integration
# kubernetes>=27.2.0

# Optional: CI/CD integration
# jenkinsapi>=0.3.12
# github3.py>=4.0.0

# Optional: Code quality
# bandit>=1.7.0
# safety>=2.3.0
# pre-commit>=3.4.0

# Optional: Performance profiling
# line-profiler>=4.1.0
# memory-profiler>=0.61.0
# cprofilev>=1.0.7

# Optional: Jupyter extensions
# jupyter-contrib-nbextensions>=0.7.0
# jupyterlab-git>=0.42.0
# jupyterlab-code-formatter>=2.2.0

# Optional: IDE integration
# jupyter-vscode-proxy>=0.1.0
# jupyterlab-lsp>=5.0.0

# Version constraints for compatibility
torch>=2.0.0,<2.1.0  # Stable version
numpy>=1.21.0,<1.25.0  # Compatible with older systems
pandas>=1.5.0,<2.0.0  # Stable pandas version

# GPU-specific packages (uncomment if using GPU)
# torch>=2.0.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# torchvision>=0.15.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# torchaudio>=2.0.0+cu118 --index-url https://download.pytorch.org/whl/cu118

# Development dependencies (uncomment for development)
# -e .
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.1.0
# black>=23.7.0
# isort>=5.12.0
# flake8>=6.1.0
# mypy>=1.5.0
# pre-commit>=3.4.0

# Production dependencies (minimal set for production)
# numpy>=1.21.0
# pandas>=1.5.0
# torch>=2.0.0
# yfinance>=0.2.0
# alpaca-trade-api>=3.0.0
# gymnasium>=0.29.0
# scikit-learn>=1.3.0
# matplotlib>=3.7.0
# fastapi>=0.100.0
# uvicorn>=0.23.0
# python-dotenv>=1.0.0
# rich>=13.5.0
# tensorboard>=2.14.0
# psutil>=5.9.0